+++ sofa-boot-project/sofa-boot-core/test-sofa-boot/src/main/java/com/alipay/sofa/test/model/stub/Stub.java
@@ -14,16 +14,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.testing;
+package com.alipay.sofa.test.model.stub;
 
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
+import org.springframework.test.context.TestContext;
+
+import java.util.Set;
 
 /**
  * @author pengym
- * @version SofaTestAutoConfiguration.java, v 0.1 2023年08月07日 15:45 pengym
+ * @version Stub.java, v 0.1 2023年08月07日 17:52 pengym
  */
-@Configuration
-@ComponentScan(basePackages = "com.alipay.sofa.testing")
-public class SofaBootTestAutoConfiguration {
+public interface Stub {
+    /**
+     * Resolve the actual stub targets
+     * @param testContext TestContext
+     * @return The actual stub targets
+     */
+    Set<Object> resolveTargets(TestContext testContext);
 }
