+++ sofa-boot-tests/sofa-boot-smoke-tests/sofa-boot-smoke-tests-integration-test/src/test/java/com/alipay/sofa/smoke/tests/integration/test/MockSpringBeanTest.java
@@ -18,7 +18,7 @@ package com.alipay.sofa.smoke.tests.integration.test;
 
 import com.alipay.sofa.runtime.api.annotation.SofaReference;
 import com.alipay.sofa.smoke.tests.integration.test.base.SofaIntegrationTestBaseCase;
-import com.alipay.sofa.testing.api.annotation.SofaMockBeanFor;
+import com.alipay.sofa.test.api.annotation.SofaMockBeanFor;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
